Uploaded assignment files for the Udemy class
There are many files that contain different classes and only one file, called "main", that calls all the other functions from the other classes. Some I skipped because they might be too simple or ask for user input. Please, feel free to check the files and share the ideas for how to make the code better.